[Backend] migrate products and shopping cart without working well, this will need be check
