+++ black-friday-sale-collection/products/getProducts.bru
@@ -5,7 +5,7 @@ meta {
 }
 
 get {
-  url: http://172.18.0.1:8080/api/v1/products/
+  url: http://localhost:8080/api/v1/products/
   body: none
   auth: none
 }
