+++ services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/application/ECommerceBlackFridaySaleApplication.java
@@ -1,0 +1,13 @@
+package org.ecommerce.blackfriday.application;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ECommerceBlackFridaySaleApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ECommerceBlackFridaySaleApplication.class, args);
+	}
+
+}
