+++ services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/application/redis/ClusterConfigurationProperties.java
@@ -1,0 +1,30 @@
+package org.ecommerce.blackfriday.application.redis;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
+
+@Component
+@ConfigurationProperties(prefix = "spring.redis.cluster")
+public class ClusterConfigurationProperties {
+
+
+    List<String> nodes;
+
+    /**
+     * Get initial collection of known cluster nodes in format {@code host:port}.
+     *
+     * @return
+     */
+
+    public List<String> getNodes() {
+        return nodes;
+    }
+
+    public void setNodes(List<String> nodes) {
+        this.nodes = nodes;
+    }
+
+
+}
