--- services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/e_commerce_black_friday_sale/config/OpenAPIConfiguration.java
@@ -1,33 +1,0 @@
-package org.ecommerce.blackfriday.e_commerce_black_friday_sale.config;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Contact;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.servers.Server;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import java.util.List;
-
-@Configuration
-public class OpenAPIConfiguration {
-
-    @Bean
-    public OpenAPI defineOpenAPI () {
-        Server server = new Server();
-        server.setUrl("http://localhost:8080");
-        server.description("Development");
-
-        Contact contact = new Contact();
-        contact.name("Hans Arias");
-
-        Info info = new Info()
-                .title("E-commerce Black Friday Sale System")
-                .version("1.0.1")
-                .description("This API exposes endpoints about an e-commerce which has as goal to replicate an event called black friday sale")
-                .contact(contact);
-
-        return new OpenAPI().info(info).servers(List.of(server));
-    }
-
-}
