--- services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/e_commerce_black_friday_sale/exceptions/ApiError.java
@@ -1,4 +1,0 @@
-package org.ecommerce.blackfriday.e_commerce_black_friday_sale.exceptions;
-
-public record ApiError(int status, String message) {
-}
