--- services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/e_commerce_black_friday_sale/exceptions/CartItemNotFoundException.java
@@ -1,8 +1,0 @@
-package org.ecommerce.blackfriday.e_commerce_black_friday_sale.exceptions;
-
-public class CartItemNotFoundException extends RuntimeException {
-
-    public CartItemNotFoundException(String cartId, String itemId) {
-        super("Cart with the [ID] " + cartId + " does not have this item added with the [ID] " + itemId);
-    }
-}
