--- services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/e_commerce_black_friday_sale/redis/ClusterConfigurationProperties.java
@@ -1,30 +1,0 @@
-package org.ecommerce.blackfriday.e_commerce_black_friday_sale.redis;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-import java.util.List;
-
-@Component
-@ConfigurationProperties(prefix = "spring.redis.cluster")
-public class ClusterConfigurationProperties {
-
-
-    List<String> nodes;
-
-    /**
-     * Get initial collection of known cluster nodes in format {@code host:port}.
-     *
-     * @return
-     */
-
-    public List<String> getNodes() {
-        return nodes;
-    }
-
-    public void setNodes(List<String> nodes) {
-        this.nodes = nodes;
-    }
-
-
-}
