--- services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/e_commerce_black_friday_sale/serv/ProductService.java
@@ -1,16 +1,0 @@
-package org.ecommerce.blackfriday.e_commerce_black_friday_sale.serv;
-
-import org.ecommerce.blackfriday.e_commerce_black_friday_sale.ent.Product;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface ProductService {
-
-    List<Product> getProducts ();
-
-    Product saveProduct (Product entity);
-
-    Optional<Product> getProduct (Long id);
-
-}
