--- services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/e_commerce_black_friday_sale/serv/ShoppingCartRedisService.java
@@ -1,10 +1,0 @@
-package org.ecommerce.blackfriday.e_commerce_black_friday_sale.serv;
-
-import org.ecommerce.blackfriday.e_commerce_black_friday_sale.dto.ShoppingCartDTO;
-
-public interface ShoppingCartRedisService {
-
-    void saveCart (ShoppingCartDTO cartDTO);
-    ShoppingCartDTO getCart (String customerId);
-
-}
