+++ services/e-commerce-black-friday-sale/src/main/java/org/ecommerce/blackfriday/shopping/cart/serv/ShoppingCartRedisService.java
@@ -1,0 +1,10 @@
+package org.ecommerce.blackfriday.shopping.cart.serv;
+
+import org.ecommerce.blackfriday.shopping.cart.dto.ShoppingCartDTO;
+
+public interface ShoppingCartRedisService {
+
+    void saveCart (ShoppingCartDTO cartDTO);
+    ShoppingCartDTO getCart (String customerId);
+
+}
